+++ src/main/kotlin/com/enterscript/noX3LanguagePlugin/language/NOX3Language.kt
@@ -2,6 +2,12 @@ package com.enterscript.noX3LanguagePlugin.language
 
 import com.intellij.lang.Language
 
-class  NOX3Language{
-    companion object : Language("X3")
+class  NOX3Language private constructor() : Language("X3") {
+    companion object {
+        @JvmField
+        val INSTANCE  = NOX3Language()
+    }
+
 }
+
+