+++ src/main/kotlin/com/enterscript/noX3LanguagePlugin/language/NOX3LexerAdapter.kt
@@ -1,6 +1,0 @@
-package com.enterscript.noX3LanguagePlugin.language
-
-import com.intellij.lexer.FlexAdapter
-
-
-class NOX3LexerAdapter : FlexAdapter(NOX3Lexer(null))