+++ src/main/kotlin/com/enterscript/noX3LanguagePlugin/language/lexer/NOX3LexerAdapter.kt
@@ -1,0 +1,8 @@
+package com.enterscript.noX3LanguagePlugin.language.lexer
+
+import com.intellij.lexer.FlexAdapter
+
+
+class NOX3LexerAdapter : FlexAdapter(NOX3Lexer(null))
+
+