+++ src/main/kotlin/com/enterscript/noX3LanguagePlugin/language/parser/NOX3ParserDefinition.kt
@@ -1,13 +1,13 @@
 package com.enterscript.noX3LanguagePlugin.language.parser
 
-import com.enterscript.noX3LanguagePlugin.language.NOX3File
 import com.enterscript.noX3LanguagePlugin.language.NOX3Language
-import com.enterscript.noX3LanguagePlugin.language.NOX3LexerAdapter
+import com.enterscript.noX3LanguagePlugin.language.lexer.NOX3LexerAdapter
+import com.enterscript.noX3LanguagePlugin.language.psi.NOX3File
 import com.enterscript.noX3LanguagePlugin.language.psi.NOX3Types
+import com.enterscript.noX3LanguagePlugin.language.psi.impl.parser.NOX3Parser
 import com.intellij.lang.ASTNode
 import com.intellij.lang.ParserDefinition
 import com.intellij.lang.PsiParser
-import com.intellij.lexer.Lexer
 import com.intellij.openapi.project.Project
 import com.intellij.psi.FileViewProvider
 import com.intellij.psi.PsiElement
@@ -15,49 +15,42 @@ import com.intellij.psi.PsiFile
 import com.intellij.psi.TokenType
 import com.intellij.psi.tree.IFileElementType
 import com.intellij.psi.tree.TokenSet
+import org.jetbrains.annotations.NotNull
 
 class NOX3ParserDefinition : ParserDefinition {
+    companion object{
+        val WHITE_SPACES = TokenSet.create(TokenType.WHITE_SPACE)
+        val COMMENTS = TokenSet.create(NOX3Types.COMMENT)
+        val FILE = IFileElementType(NOX3Language.INSTANCE)
 
-    private  companion object {
-        private val WHITE_SPACES = TokenSet.create(TokenType.WHITE_SPACE)
-        private val COMMENTS = TokenSet.create(NOX3Types.COMMENT)
-        private val FILE = IFileElementType("X3Lang", NOX3Language)
     }
 
-    override fun createLexer(project: Project): Lexer {
-        return NOX3LexerAdapter()
+    @NotNull
+    override fun createLexer(project: Project) = NOX3LexerAdapter()
+
+    @NotNull
+    override fun getWhitespaceTokens(): TokenSet =  WHITE_SPACES
+
+    @NotNull
+    override fun getCommentTokens(): TokenSet = COMMENTS
+    @NotNull
+    override fun getStringLiteralElements(): TokenSet = TokenSet.EMPTY
+
+    @NotNull
+    override fun createParser(project: Project): PsiParser =
+        NOX3Parser()
+
+    @NotNull
+    override fun getFileNodeType(): IFileElementType  = FILE
+
+    override fun createFile(viewProvider: FileViewProvider): PsiFile = NOX3File(viewProvider)
+
+    override fun spaceExistanceTypeBetweenTokens(first: ASTNode, second: ASTNode): ParserDefinition.SpaceRequirements = ParserDefinition.SpaceRequirements.MAY
+
+    @NotNull
+    override fun createElement(node: ASTNode?): PsiElement {
+        return NOX3Types.Factory.createElement(node)
     }
-
-    override fun getWhitespaceTokens(): TokenSet {
-        return WHITE_SPACES
-    }
-
-    override fun getCommentTokens(): TokenSet {
-        return COMMENTS
-    }
-
-    override fun getStringLiteralElements(): TokenSet {
-        return TokenSet.EMPTY
-    }
-
-    override fun createParser(project: Project): PsiParser {
-        return NOX3Parser()
-    }
-
-    override fun getFileNodeType(): IFileElementType {
-        return FILE
-    }
-
-    override fun createFile(viewProvider: FileViewProvider): PsiFile {
-        return NOX3File(viewProvider)
-    }
-
-    override fun spaceExistenceTypeBetweenTokens(left: ASTNode, right: ASTNode): ParserDefinition.SpaceRequirements {
-        return ParserDefinition.SpaceRequirements.MAY
-    }
-
-    override fun createElement(node: ASTNode): PsiElement = throw UnsupportedOperationException()
-