+++ src/main/kotlin/com/enterscript/noX3LanguagePlugin/language/psi/NOX3File.kt
@@ -6,10 +6,11 @@ import com.intellij.extapi.psi.PsiFileBase
 import com.intellij.openapi.fileTypes.FileType
 import com.intellij.psi.FileViewProvider
 
-
+import org.jetbrains.annotations.NotNull;
 
 class NOX3File(viewProvider: FileViewProvider) :
-    PsiFileBase(viewProvider, NOX3Language) {
+    PsiFileBase(viewProvider, NOX3Language.INSTANCE) {
+    @NotNull
     override fun getFileType(): FileType {
         return NOX3FileType.INSTANCE
     }